+++ flutter_audio_games/lib/flutter_audio_games.dart
@@ -46,4 +46,5 @@ export 'src/widgets/sounds/music/maybe_music.dart';
 export 'src/widgets/sounds/music/music.dart';
 export 'src/widgets/sounds/play_sound.dart';
 export 'src/widgets/sounds/play_sound_semantics.dart';
+export 'src/widgets/sounds/play_sounds_semantics.dart';
 export 'src/widgets/volume_list_tile.dart';